+++ R/aaa4_tinycodet_dry.R
@@ -4,14 +4,12 @@
 #' "Don't Repeat Yourself", sometimes abbreviated as "DRY",
 #' is the coding principle not to write unnecessarily repetitive code.
 #' To help in that effort, the 'tinycodet' R-package
-#' introduces a few functions:
+#' introduces a few features:
 #'
 #'
-#'  * The \link{transform_if} function
-#'  * \link[=as_int]{Atomic type casting with names and dimensions preserved}.
-#'  * The \link[=%unreal =%]{subset_if operators and the in-place unreal modifier operator}.
-#'  * The \link[=%:=%]{general in-place (mathematical) modification operator}.
-#'
+#'  * The \link{transform_if} function.
+#'  * \link[=matrix_ops]{Operators} for short-hand re-ordering matrices Row- or Column-wise.
+#'  
 #'
 #' @seealso \link{tinycodet_help}
 #'
@@ -27,11 +25,9 @@
 #'     object > 0,  log(object), object^2
 #'   )
 #' )
-#' mtcars$mpg[mtcars$cyl>6] <- (mtcars$mpg[mtcars$cyl>6])^2 # long
 #'
 #' # with tinycodet:
 #' object |> transform_if(\(x) x > 0, log, \(x) x^2, \(x) -Inf) # compact & no warning
-#' mtcars$mpg[mtcars$cyl > 6] %:=% \(x) x^2 # short
 #'
 #'