+++ R/aaa5_tinycodet_misc.R
@@ -5,7 +5,6 @@
 #' Some additional functions provided by the 'tinycodet' R-package:
 #'
 #'  * \link[=%xor%]{Infix logical operators} for exclusive-or, not-and, not-in, number-type, and string-type.
-#'  * \link[=%row~%]{Infix operators for row- and column-wise re-ordering of matrices}.
 #'  * \link[=report_inops]{Report infix operators present in the current environment, or a specified environment.}
 #'  * \link{source_selection} to source only selected objects. \cr \cr
 #'