+++ R/matrix_ops.R
@@ -9,48 +9,22 @@
 #' \cr
 #' The \code{x %col~% mat} operator re-orders the elements of every column,
 #' each column ordered independently from the other columns, of matrix \code{x},
-#' according to the ordering ranks given in matrix \code{mat}. \cr \cr
+#' according to the ordering ranks given in matrix \code{mat}. \cr
+#' \cr
+#' Note that these operators strip all attributes,
+#' except dimensions. \cr
+#' \cr
 #' 
 #'
 #' @param x a matrix
-#' @param mat a matrix with the same dimensions as \code{x},
-#' giving the ordering ranks of every element of matrix \code{x}. \cr
-#'
-#' @details
-#' If matrix \code{x} is a numeric matrix,
-#' and one wants to sort the elements of every row or column numerically,
-#' \code{x %row~% x} or \code{x %col~% x} would suffice, respectively. \cr
-#' \cr
-#' If matrix \code{x} is not numeric,
-#' sorting the elements using \code{x %row~% x} and \code{x %col~% x} is still possible,
-#' but probably not the best option.
-#' In the non-numeric case,
-#' providing a matrix of ordering ranks for \code{mat} would be faster and give more accurate ordering.
-#' See the examples section. \cr
-#' \cr
-#' If \code{mat} is a matrix of non-repeating random integers, i.e. \cr
-#' \code{mat <- sample(seq_along(x)) |> matrix(ncol = ncol(x))}) \cr
-#' then the code \cr
-#' \code{x %row~% mat} \cr
-#' will randomly shuffle the elements of every row of \code{x},
-#' where the shuffling order in each row is independent from the shuffling order in the other rows. \cr
-#' Similarly, \cr
-#' \code{x %col~% mat} \cr
-#' will randomly shuffle the elements of every column of \code{x},
-#' where the shuffling order in each column is independent from the shuffling order in the other columns. \cr
-#' \cr
-#' Re-ordering/sorting every row/column of a matrix with these operators
-#' is generally faster than doing so through loops or apply-like functions. \cr
-#' \cr
-#' Note that these operators strip all attributes except dimensions. \cr
-#' \cr
+#' @param mat a numeric matrix with the same dimensions as \code{x},
+#' giving the new ordering ranks for every element of matrix \code{x}. \cr \cr
 #' 
 #' @returns
-#' A modified matrix.
+#' A re-ordered matrix.
 #'
 #'
-#' @seealso \link{tinycodet_misc}
-#'
+#' @seealso \link{tinycodet_dry}
 #' @example inst/examples/matrix_ops.R
 #' 
 
@@ -60,22 +34,46 @@ NULL
 #' @rdname matrix_ops
 #' @export
 `%row~%` <- function(x, mat) {
-  x <- matrix(
+  .matrix_ops_checks(x, mat, sys.call())
+  out <- matrix(
     x[order(row(x), mat, decreasing = FALSE)],
     nrow=nrow(x),
     byrow = TRUE
   )
-  return(x)
+  return(out)
 }
+
 
 #' @rdname matrix_ops
 #' @export
 `%col~%` <- function(x, mat) {
-  x <- matrix(
+  .matrix_ops_checks(x, mat, sys.call())
+  out <- matrix(
     x[order(col(x), mat, decreasing = FALSE)],
     ncol=ncol(x),
     byrow = FALSE
   )
-  return(x)
+  return(out)
 }
 
+
+#' @keywords internal
+#' @noRd
+.matrix_ops_checks <- function(x, mat, abortcall) {
+  if(!is.matrix(x) || !is.matrix(mat)) {
+    stop(simpleError("both arguments must be matrices", call = abortcall))
+  }
+  if(.ndim(x) != .ndim(mat) || any(dim(x) != dim(mat))) {
+    stop(simpleError("non-conformable matrices", call = abortcall))
+  }
+  if(!is.numeric(mat)) {
+    stop(simpleError("right-hand side must be a numeric matrix of order rankings", call = abortcall))
+  }
+}
+
+
+#' @keywords internal
+#' @noRd
+.ndim <- function(x) {
+  length(dim(x))
+}