+++ NEWS.md
@@ -1,3 +1,13 @@
+
+# tinycodet 0.5.8 (Under Development)
+* **Superseded:** The `as_*` functions are superseded by the functions of the same name in the 'broadcast' package, and will be removed from 'tinycodet' in a later version.
+* **Removed:** The infix operators from the DRY category were removed.
+* **Behaviour change:** The matrix operators now only accept a numeric (integer or double) matrix on the right hand side with the same dimensions as the left hand side.
+* **Performance Improvement:** The logical operators are now slightly faster, as they now use R's own NA checks
+* Moved the matrix operators to the DRY category.
+* Simplified the matrix operators help page.
+
+
 # tinycodet 0.5.6
 * **Behaviour change:** The import functions no longer import `.Primitive` or `.Internal` functions, to comply with CRAN.
 * **Behaviour change:** `import_as()` now allows hidden (i.e. dot-prefixed) names for `alias` objects.